+++ ui/src/app/components/cookie-banner.tsx
@@ -3,7 +3,8 @@ import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
 import {StyledAnchorTag} from 'app/components/buttons';
 import {FlexRow} from 'app/components/flex';
 import colors from 'app/styles/colors';
-import {cookiesEnabled, reactStyles} from 'app/utils';
+import {reactStyles} from 'app/utils';
+import {cookiesEnabled} from 'app/utils/cookies';
 import * as React from 'react';
 
 const styles = reactStyles({
