+++ ui/src/app/pages/signed-in/nav-bar.tsx
@@ -6,9 +6,10 @@ import {SideNav} from 'app/components/side-nav';
 import {StatusAlertBanner} from 'app/components/status-alert-banner';
 import {statusAlertApi} from 'app/services/swagger-fetch-clients';
 import colors from 'app/styles/colors';
-import {cookiesEnabled, reactStyles, ReactWrapperBase, withUserProfile} from 'app/utils';
+import {reactStyles, ReactWrapperBase, withUserProfile} from 'app/utils';
 import {Profile} from 'generated/fetch';
 import * as React from 'react';
+import {cookiesEnabled} from "app/utils/cookies";
 
 const styles = reactStyles({
   headerContainer: {
