+++ ui/src/app/utils/index.tsx
@@ -24,9 +24,6 @@ const {useEffect, useState} = React;
 
 export const WINDOW_REF = 'window-ref';
 
-// Local storage key to override the API base path.
-export const LOCAL_STORAGE_API_OVERRIDE_KEY = 'allOfUsApiUrlOverride';
-
 export function isBlank(toTest: String): boolean {
   if (toTest === null || toTest === undefined) {
     return true;
@@ -132,32 +129,6 @@ export const nextSort = ({field, direction}, newField) => {
     {field, direction: direction === 'asc' ? 'desc' : 'asc'} :
     {field: newField, direction: 'asc'};
 };
-
-/**
- * See feature-detects/cookies.js in https://github.com/Modernizr
- *
- * navigator.cookieEnabled cannot detect custom or nuanced cookie blocking
- * configurations. For example, when blocking cookies via the Advanced
- * Privacy Settings in IE9, it always returns true. And there have been
- * issues in the past with site-specific exceptions.
- * Don't rely on it.
- *
- * try..catch because some in situations `document.cookie` is exposed but throws a
- * SecurityError if you try to access it; e.g. documents created from data URIs
- * or in sandboxed iframes (depending on flags/context)
- */
-export function cookiesEnabled(): boolean {
-  try {
-    // Create cookie
-    document.cookie = 'cookietest=1';
-    const ret = document.cookie.indexOf('cookietest=') !== -1;
-    // Delete cookie
-    document.cookie = 'cookietest=1; expires=Thu, 01-Jan-1970 00:00:01 GMT';
-    return ret;
-  } catch (e) {
-    return false;
-  }
-}
 
 type ReactStyles<T> = {
   readonly [P in keyof T]: React.CSSProperties;
